refactor: 쿠폰 관리 분리
